demos/test-seeds/requirements.txt
# General
apache-airflow>=2.10.0,<3
apache-airflow-providers-postgres>=5.0.0
click>=8.0
jinja2>=3.0
pandas>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# Data Quality
soda-core>=4.0.0
soda-postgres>=4.0.0

# Tests
pytest>=7.4.0

# Distributed processing (chosen at init)
pyspark>=3.5.0
